requirements.txt
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-heroku==0.3.1
gunicorn==20.0.4
Jinja2==2.11.3
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
asgiref==3.3.4
SQLAlchemy==1.3.23
sqlparse==0.4.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0

django-rest-framework==0.1.0
djangorestframework==3.12.4
